simulator/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-socketio==5.11.1
aiohttp==3.9.3
python-multipart==0.0.9
websockets==12.0 